pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opportunity-radar"
version = "0.1.0"
description = "Automated opportunity discovery and digest system"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "html2text>=2024.2.26",
    # Web UI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
opportunity-radar = "opportunity_radar.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/opportunity_radar"]
